add a skipLookupOnRetry request opt
This allows ringpop-replicator to disable key lookups on retries, and hence fixes the following issue:
```
Before retrying, ringpop looks up the destination for the given keys again, and if it finds the keys are now mapped to more than one destination, ringpop will return the above error. This behavior is not compatible with replication because the lookup during retries does not account for the replicas destinations.
```