Use a DiscoverProvider-function to discover the bootstrap hosts (#253)
By introducing the concept of DiscoverProviders in ringpop-node, we make it possible in the future to reload the host list. This is a requirement for partition healing.

When the number of hosts or their ip/ports are more dynamic (e.g. when bootstrapping over hyperbahn) it's going to be useful to reload the hosts after bootstrapping.