Deprecate "admin/reload"-endpoint (#255)
The endpoint isn't used by ringpop-admin and, according to our internal metrics, has never been called in production.
Triggering a reload was probably useful during testing when the bootstrap file changed. With partition healing, we achieve a similar result in that the discover provider will be re-queried, except that it's automatic.