Partition healing (#264)
Add the ability to heal from partitions (e.g. after a network partition). This commit adds the algorithm that checks for potential partitions by comparing reachable nodes with the bootstrip list (through the discover provider). It selects a currently non-reachable or unknown node and tries to get it's member list. Instead of merging the two memberships blindly, we detect which currently alive nodes would become faulty and instead of marking them faulty, we mark them as suspect so they have a chance to reincarnate. It also generate those changes for the target and send them over. If they're are no changes required to be able to merge the two membership, the memberships are merged and the partition is healed. More info about the algorithm is available in uber/ringpop-common.

Currently the heal algorithm is manually triggered through and admin-endpoint; a later commit will add the ability to periodically run it.