Emit stats from ringpop-test program
In preparation for automated performance and failure testing we want to be able to emit stats to a file or a network address from testpop.
This commit adds a new flag to `main.js` to control if and where stats should be emitted; if the flag is omitted, no stats are emitted.