Add support local labels to the membership
This commit adds support for labels to the local membership. To make testing cleaner/easier, I extended the `testRingpop`-function to support subtests and testing on a ringpop that isn't ready yet. I also had to add a private function - `_newIncarnationNumber` - to `Membership` to generate a new incarnation number. Otherwise the tests were flaky because of the issue descibred in #298.

Note: labels are not yet gossiped. I will add that in a separate PR.