[Feature] Key consistent forwarding (#310)
* Add option to enforce key consistency on forwarding

By default key consistency is turned of to not change functional behavior.
New logs and stats will be emitted if the keys are not consistent on
forwarded requests, but the request is still handled. This should be a
replacement for `enforceConsistency` which is a stricter consistency check
which only passes if the checksums of the ring are the same, this check is
fine with different ring checksums as long as both sides agree on the
ownership of the keys for this forwarded requests.

still needs tests to validate behavior but there were no tests for
`ensureConsistency` to start from. Both will be added in later commits.

* Add tests

* Remove key from logs to prevent data leakage through logs

* Added extra documentation to explain the consistency checks

* minor revision on the comments