Fix: Initial transition timers (#313)
This commit fixes an issue where transition timers (`faulty` -> `tombstone`) were not scheduled correctly if the membership during a `join`-response contains a `faulty`-member. Instead of duplicating the code that schedules the timers, a method is extracted from the gossip handler and used on the `set`-event (emitted after the initial join).